chore: login page migration
